change mind about underscores in quickbin_core module
Previously all C-layer binning functions had underscores on
their names, butthe quickbin.quickbin_core module itself did not.
This is not really correct. They are public members of a
private module, not private members of a public module.
Change this, and append "_handler" to the names of all functions in
quickbin._binning_handlers for disambiguation (and clarity).