replace x / y axis identifiers with i / j axis identifiers
When writing this, I followed the `binned_statistic_2d()` convention of
giving the first two arguments "x"  and "y" identifiers and mapping them
to the first and second axes respectively. However, because numpy uses
row-major axis order, this maps arguments named "x" to the y axis and
arguments named "y" to the x axis. Using ill-named "x" and "y" axis
identifiers like is a common but not universal convention in scientific
Python. It confused me for the 100th time earlier today, and, in the
context of `quickbin`, it has confused me for the last time. I have replaced
all "x" and "y" axis identifiers with "i" and "j" in order to emphasize
(1) that they map to the first and second axes respectively, and (2) that
both we and our friend numpy are using row-major indexing.

This is a purely cosmetic refactor. No behavior has changed.